Restructered _process.py file, fixed platform compatibility issues, enhanced plt command.
Description:
	- Replaced if/else checks for platform with state map
	- Separated platform specific calls to respective functions
	- Added showing default platform to plt command
		* `clvm plt` without arguments will print the current default platform

+++ src/pyclvm/plt.py
@@ -16,7 +16,6 @@ def _get_cache_path():
 
 def _create_cache() -> None:
     """creates the cache directory and file if they doesn't exist"""
-
     makedirs(
         name=path.join(Path.home(), ".clvm"),
         mode=0o700,
@@ -30,13 +29,17 @@ def _create_cache() -> None:
 
 def _set_default_platform(platform: str) -> None:
     """sets the default platform"""
-    with open(_get_cache_path(), "w") as cache:
-        data = {}
-        data["platform"] = f"{platform.upper()}"
-        cache.seek(0)
-        json.dump(data, cache, indent=4)
-        cache.truncate()
-    print(f"Default platform is {platform.upper()}")
+    supported_platforms = {"AWS", "GCP", "AZURE"}
+    if platform in supported_platforms:
+        with open(_get_cache_path(), "w") as cache:
+            data = {}
+            data["platform"] = f"{platform.upper()}"
+            cache.seek(0)
+            json.dump(data, cache, indent=4)
+            cache.truncate()
+        print(f"Default platform is {platform.upper()}")
+    else:
+        _unsupported_platform(platform)
 
 
 def _default_platform(**kwargs) -> Union[str, Any]:
@@ -48,16 +51,18 @@ def _default_platform(**kwargs) -> Union[str, Any]:
     if "platform" not in kwargs.keys():
         with open(_get_cache_path(), "r") as cache:
             platform_ = json.load(cache)
-        return platform_["platform"]
+        return platform_["platform"], supported_platforms
     else:
         platform_ = str(kwargs["platform"]).upper()
         if platform_ in supported_platforms:
-            return platform_.upper()
+            return platform_.upper(), supported_platforms
         else:
             _unsupported_platform(platform_)
 
 
 def _unsupported_platform(platform: Union[str, None]) -> None:
+    if platform:
+        platform = platform.upper()
     print(
         "Unsupported platform!",
         f"{platform} is not in the supported platforms list",
@@ -66,18 +71,24 @@ def _unsupported_platform(platform: Union[str, None]) -> None:
     )
 
 
-def plt(platform: str):
+def plt(*platform: str, **kwargs: str) -> Union[str, None]:
     """
     change default platform (AWS, GCP, AZURE)
 
     Args:
-        platform (str): one of the 3 platforms (AWS, GCP, AZURE)
+        **platform (str): one of the 3 platforms (AWS, GCP, AZURE). Default is AWS.
 
     Returns:
         None
     """
-    supported_platforms = {"AWS", "GCP", "AZURE"}
-    if platform.upper() in supported_platforms:
-        _set_default_platform(platform)
+    default_platform, supported_platforms = _default_platform(**kwargs)
+    if platform:
+        platform_name = platform[0].upper()
     else:
-        _unsupported_platform(platform)
+        print(f"Default platform is {default_platform}")
+        return
+
+    if platform_name in supported_platforms:
+        _set_default_platform(platform_name)
+    else:
+        _unsupported_platform(platform_name)
